+++ app/src/main/java/com/example/playlistmaker/data/settings/SettingsRepository.kt
@@ -5,4 +5,5 @@ import com.example.playlistmaker.domain.settings.model.ThemeSettings
 interface SettingsRepository {
     fun getThemeSettings(): ThemeSettings
     fun updateThemeSetting(settings: ThemeSettings)
+    fun switchTheme(darkThemeEnabled: Boolean)
 }
