+++ app/src/main/java/com/example/playlistmaker/data/settings/theme/ThemeRepository.kt
@@ -2,10 +2,11 @@ package com.example.playlistmaker.data.settings.theme
 
 import android.content.Context
 import android.content.SharedPreferences
+import androidx.appcompat.app.AppCompatDelegate
 import com.example.playlistmaker.domain.settings.model.ThemeSettings
 
 class ThemeRepository(private val context: Context) {
-
+    var darkTheme = false
     companion object {
         private const val PREFS_NAME = "ThemePrefs"
         private const val KEY_DARK_THEME = "isDarkTheme"
@@ -28,5 +29,15 @@ class ThemeRepository(private val context: Context) {
         val editor = sharedPreferences.edit()
         editor.putBoolean(KEY_DARK_THEME, isDarkTheme)
         editor.apply()
+    }
+    fun switchTheme(darkThemeEnabled: Boolean) {
+        darkTheme = darkThemeEnabled
+        AppCompatDelegate.setDefaultNightMode(
+            if (darkThemeEnabled) {
+                AppCompatDelegate.MODE_NIGHT_YES
+            } else {
+                AppCompatDelegate.MODE_NIGHT_NO
+            }
+        )
     }
 }
