+++ app/src/main/java/com/example/playlistmaker/domain/settings/SettingsInteractor.kt
@@ -5,4 +5,5 @@ import com.example.playlistmaker.domain.settings.model.ThemeSettings
 interface SettingsInteractor {
     fun getThemeSettings(): ThemeSettings
     fun updateThemeSetting(settings: ThemeSettings)
+    fun switchTheme(darkThemeEnabled: Boolean)
 }
