+++ app/src/main/java/com/example/playlistmaker/domain/settings/impl/SettingsInteractorImpl.kt
@@ -14,4 +14,8 @@ class SettingsInteractorImpl(
     override fun updateThemeSetting(settings: ThemeSettings) {
         settingsRepository.updateThemeSetting(settings)
     }
+
+    override fun switchTheme(darkThemeEnabled: Boolean) {
+        settingsRepository.switchTheme(darkThemeEnabled)
+    }
 }
