+++ app/src/main/res/values-night/themes.xml
@@ -93,6 +93,10 @@
         <item name="cornerRadius">54dp</item>
         <item name="android:textStyle">bold</item>
     </style>
+    <style name="TabLayout">
+        <item name="tabIndicatorColor">@color/white</item>
+        <item name="tabTextColor">@color/white</item>
+    </style>
 
 </resources>
