+++ app/src/main/res/values/themes.xml
@@ -47,7 +47,7 @@
         <item name="android:layout_height">wrap_content</item>
         <item name="android:textSize">22sp</item>
         <item name="android:textStyle">bold</item>
-        <item name="android:layout_marginStart">28dp</item>
+        <item name="android:layout_marginStart">16dp</item>
         <item name="android:layout_marginTop">14dp</item>
         <item name="android:fontFamily">@font/ys_display_medium</item>
         <item name="android:textColor">@color/text_main</item>
@@ -85,9 +85,15 @@
     </style>
     <style name="button">
         <item name="android:textColor">@color/white</item>
+        <item name="android:textAllCaps">false</item>
         <item name="android:backgroundTint">@color/black</item>
         <item name="android:textSize">14sp</item>
         <item name="android:fontFamily">@font/ys_display_medium</item>
         <item name="cornerRadius">54dp</item>
     </style>
+    <style name="TabLayout">
+        <item name="tabIndicatorColor">@color/black</item>
+        <item name="tabTextColor">@color/black</item>
+    </style>
+
 </resources>
